Versão 3 - Funções em Biblioteca Separada
Precisa reorganizar as funções e corrigir erros
LOG:
Retiradas as GLOBAIS
Todas as funções que usavam as globais sofreram alterações, tanto de parâmetro como de variável 
[ - produto *p = vetor dos produtos em estoque
  - int n = variável da quantidade de produtos cadastrados até o momento
  - int z = variável responsável pelos índices ]
